requirements.txt
# LiveKit and MCP dependencies
mcp
livekit-agents>=1.0.0rc7
livekit-plugins-openai>=1.0.0rc7
livekit-plugins-silero>=1.0.0rc7
livekit-plugins-azure>=1.0.0rc7

# Azure Speech Services
azure-cognitiveservices-speech

# Web scraping and browser automation
playwright
beautifulsoup4
html5lib
lxml
lxml-html-clean
brotli

# HTTP and API clients
httpx>=0.25.2
requests
aiohttp

# Web framework
fastapi
uvicorn
pydantic

# Utilities
python-dotenv
setuptools
pytz
geopy

# Search and information retrieval
wikipedia-api
duckduckgo-search
google-search

# Enhanced search with RAG
numpy>=1.24.0
scipy>=1.10.0
